Support optional id for divider items
I added an optional id for divider items. This allows dividers to be removed by calling `Builder.remove(int)`.

Use case: You have a large menu with many dividers. You remove some menu items using the builder depending on some application state, so it could happen that two or more dividers are drawn side by side. With this PR you could remove an menu item and it's associated divider by their ids.

PS: This is *not* tested yet. But it seems so trivial that I decided to just make a quick on-page proposal. I hope, this is ok!